requirement.txt
mediapipe
opencv-python
tensorflow
pillow
fastapi
uvicorn
python-multipart
uvicorn[standard]
websockets
vosk==0.3.45
pyaudio==0.2.13